refactor: export validation function to separate file
